Finish config of MSW
